Renamed the module to k3po.launcher. Renamed the connectURI to control.
