+++ pom.xml
@@ -27,7 +27,7 @@
         <module>lang</module>
         <module>driver</module>
         <module>k3po-maven-plugin</module>
-        <module>executable</module>
+        <module>launcher</module>
     </modules>
 
     <properties>
